import added

+++ models.py
@@ -12,7 +12,7 @@ class Models:
         
     def Select_model(self,model):
         # #   wbcd (diagnostic)
-        # self.y_train = self.y_train.astype(int)
+        self.y_train = self.y_train.astype(int)
         #______________________________________
         if model == "SVM":
             y_pred=self.SVM()
